feat: created IntegralIRAddress and StringIRAddress abstract types
