+++ src/main/java/me/mtk/torrey/frontend/ir/addressing/IRConstAddress.java
@@ -3,7 +3,7 @@ package me.mtk.torrey.frontend.ir.addressing;
 /**
  * Represents an address with CONSTANT addressing mode.
  */
-public final class IRConstAddress extends IRAddress
+public final class IRConstAddress extends IntegralIRAddress
 {
   public IRConstAddress(long constant)
   {
