+++ src/main/java/me/mtk/torrey/frontend/ir/addressing/IRLabelAddress.java
@@ -3,7 +3,7 @@ package me.mtk.torrey.frontend.ir.addressing;
 /**
  * Represents an address with LABEL addressing mode.
  */
-public final class IRLabelAddress extends IRAddress
+public final class IRLabelAddress extends StringIRAddress
 {
   // The number of the label address.
   private static int num = 0;
