+++ src/main/java/me/mtk/torrey/frontend/ir/addressing/IRNameAddress.java
@@ -3,7 +3,7 @@ package me.mtk.torrey.frontend.ir.addressing;
 /**
  * Represents an address with NAME addressing mode.
  */
-public final class IRNameAddress extends IRAddress
+public final class IRNameAddress extends StringIRAddress
 {
   public IRNameAddress(String name)
   {
