+++ src/main/java/me/mtk/torrey/frontend/ir/addressing/IRNullAddress.java
@@ -4,14 +4,14 @@ package me.mtk.torrey.frontend.ir.addressing;
  * An address that has constant addressing mode and indicates
  * that an expression does not evaluate to anything
  */
-public final class IRNullAddress extends IRAddress
+public final class IRNullAddress extends IntegralIRAddress
 {
   public IRNullAddress()
   {
     super(IRAddressingMode.CONSTANT, 0);
   }
 
-  public IRAddress toIRConstantAddress()
+  public IRConstAddress toIRConstantAddress()
   {
     return new IRConstAddress(0);
   }
